ToggUltima 3.1 Permission Update
Added the Bouncer script. This script is only to disallow or allow button usage.

Notice: This is not a end all be all to world interaction security and will only keep the players who are not using mods in check. This script is just a deterrent and will only be developed as such.